# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/trips": {
      "post": {
        "operationId": "TripsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/trips/matches": {
      "get": {
        "operationId": "TripsController_matches",
        "parameters": [
          {
            "name": "destination",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "omni-dance API",
    "description": "API de la plataforma omni-dance (SBK Santiago): eventos, QR, sesiones, tickets, social, academias, gamificación, RBAC y admin.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "omnidance_session"
      }
    },
    "schemas": {
...
      "CreateTripDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}